update & move `constants`, update default `uiHighlight` to #808080
