--- src/libs/constants.ts
@@ -1,10 +1,0 @@
-export const CUSTOM_ENDPOINT_KEY = "app-portal-custom-endpoints";
-
-export const chainColours: Record<string, string> = {
-	"cennznet azalea": "#1130FF",
-	"cennznet nikau": "#05b210",
-	"cennznet rata": "#9847FF",
-	"development": "#808080",
-};
-
-export const COMMIT_HASH: string = String(process.env.COMMIT_HASH);
